third_party/blink/web_tests/fast/forms/time-multiple-fields/time-multiple-fields-stepup-stepdown-from-renderer-expected.txt

Check stepping-up and -down for time input fields from renderer. No cases of empty initial values.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Function arguments are (value, step, min, max, [keySequence]).
Normal cases
PASS stepUp("07:13", null, null, null) is "07:14"
PASS stepDown("07:13", null, null, null) is "07:12"
PASS test("07:13", null, null, null, ["ArrowUp", "ArrowUp", "ArrowUp"]) is "07:16"
PASS test("07:13", null, null, null, ["ArrowDown", "ArrowDown", "ArrowDown"]) is "07:10"
Milliseconds
PASS stepUp("07:13:00.000", 0.001, null, null) is "07:13:00.001"
PASS stepDown("07:13:00.001", 0.001, null, null) is "07:13:00.000"
PASS stepUp("07:13:00.999", 0.001, null, null) is "07:13:00.000"
PASS stepDown("07:13:00.000", 0.001, null, null) is "07:13:00.999"
PASS stepUp("07:13:00.000", 0.250, null, null) is "07:13:00.250"
PASS stepDown("07:13:00.500", 0.250, null, null) is "07:13:00.250"
PASS stepUp("07:13:00.750", 0.250, null, null) is "07:13:00.000"
PASS stepDown("07:13:00.000", 0.250, null, null) is "07:13:00.750"
PASS stepUp("07:13:00.099", 0.100, null, null) is "07:13:00.100"
PASS stepDown("07:13:00.101", 0.100, null, null) is "07:13:00.100"
PASS stepUp("07:13:00.500", 0.0001, null, null) is "07:13:00.501"
PASS stepDown("07:13:00.500", 0.0001, null, null) is "07:13:00.499"
PASS stepUp("07:13:00.500", 0, null, null) is "07:13:00.000"
PASS stepDown("07:13:00.500", 0, null, null) is "07:13:00.000"
PASS stepUp("07:13:00.500", 60, null, null) is "07:13:00.000"
PASS stepDown("07:13:00.500", 60, null, null) is "07:13:00.000"
PASS stepUp("07:13:00.500", 0.100, "00:00:00.050", null) is "07:13:00.550"
PASS stepDown("07:13:00.500", 0.100, "00:00:00.050", null) is "07:13:00.450"
PASS test("07:13:00.500", 0.100, null, null, ["Delete", "ArrowUp"]) is "07:13:00.000"
PASS test("07:13:00.500", 0.100, null, null, ["Delete", "ArrowDown"]) is "07:13:00.900"
PASS test("07:13:00.500", 0.100, "00:00:00.050", null, ["Delete", "ArrowUp"]) is "07:13:00.050"
PASS test("07:13:00.500", 0.100, "00:00:00.050", null, ["Delete", "ArrowDown"]) is "07:13:00.950"
PASS stepUp("07:13:00.010", 0.001, "07:13:00.005", "07:13:00.010") is "07:13:00.005"
PASS stepDown("07:13:00.005", 0.001, "07:13:00.005", "07:13:00.010") is "07:13:00.010"
PASS stepUp("07:13:00.005", 0.001, "07:13:00.010", "07:13:00.020") is "07:13:00.010"
PASS stepDown("07:13:00.005", 0.001, "07:13:00.010", "07:13:00.020") is "07:13:00.020"
PASS stepUp("07:13:00.005", 0.001, "07:13:00.002", "07:13:00.002") is "07:13:00.002"
PASS stepDown("07:13:00.005", 0.001, "07:13:00.002", "07:13:00.002") is "07:13:00.002"
PASS stepUp("07:13:00.035", 0.01, "07:13:00.010", "07:13:00.035") is "07:13:00.010"
PASS stepDown("07:13:00.035", 0.01, "07:13:00.010", "07:13:00.035") is "07:13:00.030"
PASS stepUp("07:13:00.010", 0.001, "07:13:00.010", "07:13:01.020") is "07:13:00.011"
PASS stepDown("07:13:00.010", 0.001, "07:13:00.010", "07:13:01.020") is "07:13:00.009"
Seconds
PASS stepUp("07:13:00", 1, null, null) is "07:13:01"
PASS stepDown("07:13:01", 1, null, null) is "07:13:00"
PASS stepUp("07:13:59", 0.001, null, null) is "07:13:00.000"
PASS stepDown("07:13:00", 0.001, null, null) is "07:13:59.000"
PASS stepUp("07:13:00", 10, null, null) is "07:13:10"
PASS stepDown("07:13:30", 10, null, null) is "07:13:20"
PASS stepUp("07:13:50", 10, null, null) is "07:13:00"
PASS stepDown("07:13:00", 10, null, null) is "07:13:50"
PASS stepUp("07:13:09", 10, null, null) is "07:13:10"
PASS stepDown("07:13:11", 10, null, null) is "07:13:10"
PASS stepUp("07:13:30", 2.5, null, null) is "07:13:31.000"
PASS stepDown("07:13:30", 2.5, null, null) is "07:13:29.000"
PASS stepUp("07:13:30", 0, null, null) is "07:13:00"
PASS stepDown("07:13:30", 0, null, null) is "07:13:00"
PASS stepUp("07:13:30", 3600, null, null) is "07:13:00"
PASS stepDown("07:13:30", 3600, null, null) is "07:13:00"
PASS stepUp("07:13:30", 11, null, null) is "07:13:31"
PASS stepDown("07:13:30", 11, null, null) is "07:13:29"
PASS stepUp("07:13:30", 10, "00:00:05", null) is "07:13:35"
PASS stepDown("07:13:30", 10, "00:00:05", null) is "07:13:25"
PASS test("07:13:30", 10, null, null, ["Delete", "ArrowUp"]) is "07:13:00"
PASS test("07:13:30", 10, null, null, ["Delete", "ArrowDown"]) is "07:13:50"
PASS test("07:13:30", 10, "00:00:05", null, ["Delete", "ArrowUp"]) is "07:13:05"
PASS test("07:13:30", 10, "00:00:05", null, ["Delete", "ArrowDown"]) is "07:13:55"
PASS stepUp("07:13:10", 1, "07:13:05", "07:13:10") is "07:13:05"
PASS stepDown("07:13:05", 1, "07:13:05", "07:13:10") is "07:13:10"
PASS stepUp("07:13:05", 1, "07:13:10", "07:13:20") is "07:13:10"
PASS stepDown("07:13:05", 1, "07:13:10", "07:13:20") is "07:13:20"
PASS stepUp("07:13:05", 1, "07:13:02", "07:13:02") is "07:13:02"
PASS stepDown("07:13:05", 1, "07:13:02", "07:13:02") is "07:13:02"
PASS stepUp("07:13:35", 10, "07:13:10", "07:13:35") is "07:13:10"
PASS stepDown("07:13:35", 10, "07:13:10", "07:13:35") is "07:13:30"
PASS stepUp("07:13:10", 1, "07:13:10", "07:14:20") is "07:13:11"
PASS stepDown("07:13:10", 1, "07:13:10", "07:14:20") is "07:13:09"
Minutes
PASS stepUp("07:13", 1, null, null) is "07:14:00"
PASS stepDown("07:13", 1, null, null) is "07:12:00"
PASS stepUp("07:59", 0.001, null, null) is "07:00:00.000"
PASS stepDown("07:00", 0.001, null, null) is "07:59:00.000"
PASS stepUp("07:00", 600, null, null) is "07:10"
PASS stepDown("07:10", 600, null, null) is "07:00"
PASS stepUp("07:50", 600, null, null) is "07:00"
PASS stepDown("07:00", 600, null, null) is "07:50"
PASS stepUp("07:13", 600, null, null) is "07:20"
PASS stepDown("07:13", 600, null, null) is "07:10"
PASS stepUp("07:13", 150, null, null) is "07:14:00"
PASS stepDown("07:13", 150, null, null) is "07:12:00"
PASS stepUp("07:13", 0, null, null) is "07:14"
PASS stepDown("07:13", 0, null, null) is "07:12"
PASS stepUp("07:13", 36000, null, null) is "07:00"
PASS stepDown("07:13", 36000, null, null) is "07:00"
PASS stepUp("07:13", 660, null, null) is "07:14"
PASS stepDown("07:13", 660, null, null) is "07:12"
PASS stepUp("07:13", 600, "00:05", null) is "07:15"
PASS stepDown("07:13", 600, "00:05", null) is "07:05"
PASS test("07:13", 600, null, null, ["Delete", "ArrowUp"]) is "07:00"
PASS test("07:13", 600, null, null, ["Delete", "ArrowDown"]) is "07:50"
PASS test("07:13", 600, "00:05", null, ["Delete", "ArrowUp"]) is "07:05"
PASS test("07:13", 600, "00:05", null, ["Delete", "ArrowDown"]) is "07:55"
PASS stepUp("07:10", 1, "07:05", "07:10") is "07:05:00"
PASS stepDown("07:05", 1, "07:05", "07:10") is "07:10:00"
PASS stepUp("07:05", 1, "07:10", "07:20") is "07:10:00"
PASS stepDown("07:05", 1, "07:10", "07:20") is "07:20:00"
PASS stepUp("07:05", 1, "07:00", "07:00") is "07:00:00"
PASS stepDown("07:05", 1, "07:00", "07:00") is "07:00:00"
PASS stepUp("07:35", 600, "07:10", "07:35") is "07:10"
PASS stepDown("07:35", 600, "07:10", "07:35") is "07:30"
PASS stepUp("07:10", 1, "07:10", "08:20") is "07:11:00"
PASS stepDown("07:10", 1, "07:10", "08:20") is "07:09:00"


PASS successfullyParsed is true

TEST COMPLETE
